refactor(crypto): move kes-cli to standalone crate (#702)
