+++ Cargo.toml
@@ -21,4 +21,5 @@ members = [
   "examples/n2n-miniprotocols",
   "examples/n2c-miniprotocols",
   "examples/p2p",
+  "kes-cli",
 ]
