--- pallas-crypto/src/kes/cli/mod.rs
@@ -1,25 +1,0 @@
-//! Module for summed kes cli
-
-/// Public key derivation
-pub mod derive_pk;
-
-/// Signing key derivation
-pub mod derive_sk;
-
-/// Seed generation
-pub mod generate_seed;
-
-/// Signing key generation
-pub mod generate_sk;
-
-/// Period of signing key
-pub mod period;
-
-/// Message signing
-pub mod sign;
-
-/// Signing key updating
-pub mod update;
-
-/// Message verifying
-pub mod verify;
